Move all pyIAST code to a single package
